created structure and added @types to the app with help yarn add: - @types/react-redux - @types/react-router-dom
