+++ package.json
@@ -12,6 +12,8 @@
     "@types/node": "^14.14.6",
     "@types/react": "^16.9.55",
     "@types/react-dom": "^16.9.9",
+    "@types/react-redux": "^7.1.10",
+    "@types/react-router-dom": "^5.1.6",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.5",
     "node-sass": "4.14.1",
