+++ src/App/App.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import logo from '../logo.svg';
+import styles from './App.module.scss';
+
+export default function App() {
+  return (
+      <BrowserRouter>
+        <div className={styles.App}>
+          <header className={styles.App_header}>
+            <img src={logo} className={styles.App_logo} alt="logo" />
+            <p>
+              Edit <code>src/App.js</code> and save to reload.
+            </p>
+            <a
+              className={styles.App_link}
+              href="https://reactjs.org"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              Learn React
+            </a>
+          </header>
+        </div>
+      </BrowserRouter>
+  );
+
+}
