+++ src/redux/rootReducer.js
@@ -1,0 +1,3 @@
+import {combineReducers} from "redux";
+
+export const rootReducer = combineReducers({});
